+++ mjml/core/registry.py
@@ -4,14 +4,45 @@ __all__ = ['components', 'register_components', 'register_core_components']
 
 from typing import List, Type
 
+
 components = {}
 
 def register_core_components():
-    from ..elements import (MjAccordion, MjAccordionElement, MjAccordionText, MjAccordionTitle,
-        MjButton, MjText, MjSection, MjCarousel, MjCarouselImage, MjColumn, MjBody,
-        MjGroup, MjHero, MjImage, MjNavbar, MjNavbarLink, MjDivider, MjSocial, MjSocialElement, MjSpacer, MjTable, MjRaw, MjWrapper)
-    from ..elements.head import (MjAttributes, MjBreakpoint, MjFont, MjHead, MjHtmlAttributes, MjPreview, MjStyle,
-        MjTitle)
+    from ..elements import (
+        MjAccordion,
+        MjAccordionElement,
+        MjAccordionText,
+        MjAccordionTitle,
+        MjBody,
+        MjButton,
+        MjCarousel,
+        MjCarouselImage,
+        MjColumn,
+        MjDivider,
+        MjGroup,
+        MjHero,
+        MjImage,
+        MjNavbar,
+        MjNavbarLink,
+        MjRaw,
+        MjSection,
+        MjSocial,
+        MjSocialElement,
+        MjSpacer,
+        MjTable,
+        MjText,
+        MjWrapper,
+    )
+    from ..elements.head import (
+        MjAttributes,
+        MjBreakpoint,
+        MjFont,
+        MjHead,
+        MjHtmlAttributes,
+        MjPreview,
+        MjStyle,
+        MjTitle,
+    )
 
     register_components([
         MjAccordion,