+++ mjml/elements/head/mj_head.py
@@ -9,4 +9,3 @@ class MjHead(HeadComponent):
 
     def handler(self):
         return self.handlerChildren()
-