+++ mjml/elements/head/mj_style.py
@@ -17,4 +17,3 @@ class MjStyle(HeadComponent):
         add = self.context['add']
         inline_attr = 'inlineStyle' if (self.get_attr('inline') == 'inline') else 'style'
         add(inline_attr, self.getContent())
-