+++ tests/mjml2html_test.py
@@ -5,7 +5,6 @@ from unittest import TestCase
 from mjml import mjml_to_html
 
 
-
 class MJML2HTMLTest(TestCase):
     def test_can_handle_comments_in_mjml(self):
         mjml = (
@@ -16,4 +15,3 @@ class MJML2HTMLTest(TestCase):
             '</mjml>'
         )
         mjml_to_html(StringIO(mjml))
-