+++ tests/upstream_alignment_test.py
@@ -3,12 +3,11 @@ from json import load as json_load
 from unittest import SkipTest, TestCase
 
 from bs4 import BeautifulSoup
-from ddt import ddt as DataDrivenTestCase, data as ddt_data
+from ddt import data as ddt_data, ddt as DataDrivenTestCase
 from htmlcompare import assert_same_html
 
 from mjml import mjml_to_html
 from mjml.testing_helpers import get_mjml_fp, load_expected_html
-
 
 
 @DataDrivenTestCase
@@ -37,6 +36,8 @@ class UpstreamAlignmentTest(TestCase):
         'mj-head-with-comment',
         'mj-image-with-empty-alt-attribute',
         'mj-image-with-href',
+        'mj-section-with-full-width',
+        'mj-section-with-css-class',
         'mj-section-with-mj-class',
         'mj-section-with-background-url',
         'mj-font',
@@ -84,7 +85,7 @@ class UpstreamAlignmentTest(TestCase):
 
     def test_can_use_css_inlining(self):
         try:
-            import css_inline
+            import css_inline  # noqa: unused-import
         except ImportError:
             raise SkipTest('"css_inline" not installed')
         test_id = 'css-inlining'
@@ -147,6 +148,5 @@ class UpstreamAlignmentTest(TestCase):
         body_actual = BeautifulSoup(result.html, 'html.parser').body
         actual_text = body_actual.get_text().strip()
         assert (expected_text == actual_text)
-        actual_html = (body_actual.select('table > tr > td > div')[0]).renderContents()
+        actual_html = (body_actual.select('.mj-column-per-100 div')[0]).renderContents()
         assert (b'foo <b>bar</b>.' == actual_html)
-